hometask bottom nav bar, viewPager ver 1.2 with fixes
