+++ app/src/main/java/com/itis2018hometask/itis2018hometask/PeopleItem.kt
@@ -1,3 +1,3 @@
 package com.itis2018hometask.itis2018hometask
 
-class PeopleItem(var name: String, var points: Int, var image: Int)
+data class PeopleItem(var name: String, var points: Int, var image: Int)
