+++ app/src/main/res/layout/people_item.xml
@@ -3,8 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginHorizontal="10dp"
-    android:layout_marginVertical="5dp">
+    android:layout_marginHorizontal="12dp"
+    android:layout_marginVertical="4dp">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
